requirements.txt

python-telegram-bot>=20.0
python-dotenv>=1.0.0
solana>=0.30.0
Flask>=2.0.0
pyngrok>=7.0.0
httpx>=0.24.0
